+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support for publisher maps for service connector
 - Support to attach and detach instances to `oci_core_instance_pool` 
 - Support for GoldenGate service
+- Support for unified agent configuration added to `oci_logging` service
 
 ## 4.16.0 (March 03, 2021)